+++ internal/iomem/iomem_others.go
@@ -1,8 +1,8 @@
 //go:build !linux
 
-// Package pagecache provides OS-specific helpers to advise the kernel
+// Package iomem provides OS-specific helpers to advise the kernel
 // about page cache usage for accessed files.
-package pagecache
+package iomem
 
 import (
 	"os"
@@ -11,7 +11,7 @@ import (
 )
 
 // HintStreaming is a no-op on non-Linux builds. It errors on a nil file to
-// match the Linux implementation's contract (callWithFd in pagecache_linux.go).
+// match the Linux implementation's contract (callWithFd in iomem_linux.go).
 func HintStreaming(f *os.File) error {
 	if f == nil {
 		return errors.New("nil file")
